+++ NEWS.md
@@ -13,6 +13,7 @@ The list below highlights major breaking changes, and please note that significa
 
 # Major changes in v0.25
 
+- Changed API to `testFactorResidualBinary(fct, meas::Tuple, (T_i, param_i),...)` to grow beyond binary.
 - PPE methods used keyword `method::AbstractPointParametricType` which is now replaced with the keyword `ppeType`.
 - Belief points are now stored as a `Vector{P}` (rather than legacy `Matrix`), and currently still under the restriction `P <: AbstractVector{<:Real}`.  Objective is moving to `P` any user desired point that fits with the JuliaManifolds/Manifolds.jl patterns.
 - Deprecating use of `ensureAllInitialized!`, use `initAll!` instead.